refactor: Change position of getting articles data logic
기존 press container에서 다루던 articles data를 상위 컴포넌트로 빼고 props로 전달
