+++ src/components/main/main__left/ns__container/NsContainerComponent.ts
@@ -1,10 +1,17 @@
-import { Props, State, ViewState } from '@custom-types/types';
+import { Article, Props, State, ViewState } from '@custom-types/types';
 import { Component } from '@custom-types/interfaces';
 import { NsContainerModel } from '@components/main/main__left/ns__container/NsContainerModel.js';
 import { NsContainerView } from '@components/main/main__left/ns__container/NsContainerView.js';
 import { NsNavbarComponent } from '@components/main/main__left/ns__container/ns__navbar/NsNavbarComponent.js';
 import { NsPressContainerComponent } from '@components/main/main__left/ns__container/ns__press-container/NsPressContainerComponent.js';
 import { NsCategoryContainerObserverViewComponent } from '@components/main/main__left/ns__container/ns__category-container/NsCategoryContainerObserverViewComponent.js';
+import { customGet } from '@utils/customFetch.js';
+import {
+  BASIC_URL,
+  PRESS_CONTAINER_ITEM_COUNT,
+  PRESS_CONTAINER_PAGE_END,
+} from '@src/constants/constants.js';
+import { pickRandomData } from '@utils/pickRandomData.js';
 
 export class NsContainerComponent implements Component {
   private _model: NsContainerModel;
@@ -12,7 +19,9 @@ export class NsContainerComponent implements Component {
   constructor(props?: Props) {
     this._model = new NsContainerModel();
     this._view = new NsContainerView();
-    this.attachChildComponents();
+    // props 처리부터 해서 child component attach하기
+    const articlesPromise = this.getRandomArticles();
+    this.attachChildComponents({ articlesPromise });
 
     const view: ViewState = 'GRID';
     const handleToView = this.handleToView.bind(this);
@@ -38,8 +47,10 @@ export class NsContainerComponent implements Component {
 
   attachChildComponents(props?: Props) {
     const nsNavbar = new NsNavbarComponent();
-    const nsPressContainer = new NsPressContainerComponent();
-    const nsCategoryContainer = new NsCategoryContainerObserverViewComponent();
+    const nsPressContainer = new NsPressContainerComponent(props);
+    const nsCategoryContainer = new NsCategoryContainerObserverViewComponent(
+      props,
+    );
 
     nsNavbar.attachTo(this);
     nsPressContainer.attachTo(this);
@@ -49,5 +60,19 @@ export class NsContainerComponent implements Component {
   handleToView(state: State) {
     const { view } = state;
     this.setState({ view });
+  }
+
+  async getArticles(): Promise<Article[]> {
+    const articleData = await customGet(`${BASIC_URL}/articles`).then((res) =>
+      res.json(),
+    );
+    return articleData;
+  }
+
+  async getRandomArticles() {
+    const articleData = await this.getArticles();
+    const totalItemCount =
+      PRESS_CONTAINER_PAGE_END * PRESS_CONTAINER_ITEM_COUNT;
+    return pickRandomData(articleData, totalItemCount);
   }
 }
